+++ R/gk.R
@@ -125,7 +125,8 @@ gk_w <- function(x,pvar,qvar,pervar,prodID, sample) {
 #' across all periods in a given window. Default is not to match.
 #' @param window An integer specifying the length of the window.
 #' @param splice the splicing method to use to extend the index. Valid methods are
-#' window, movement, half, mean, fbew or fbmw. The default is mean.
+#' window, movement, half, mean, fbew, fbmw, wisp, hasp or mean_pub. The default is mean.
+#' See details for important considerations when using fbew and fbmw.
 #' @details The splicing methods are used to update the price index when new data become
 #' available without changing prior index values. The window, movement, half and mean splices
 #' use the most recent index value as the base period, which is multiplied by a price movement
@@ -153,7 +154,7 @@ gk_w <- function(x,pvar,qvar,pervar,prodID, sample) {
 GKIndex <- function(x, pvar, qvar, pervar, prodID, sample = "", window, splice = "mean"){
 
   # check that only valid splice methods are chosen
-  if(!(tolower(splice) %in% c("mean", "window", "movement", "half", "fbew", "fbmw"))){
+  if(!(tolower(splice) %in% c("mean", "window", "movement", "half", "fbew", "fbmw", "wisp", "hasp", "mean_pub"))){
     stop("Not a valid splicing method.")
   }
 
@@ -229,6 +230,9 @@ GKIndex <- function(x, pvar, qvar, pervar, prodID, sample = "", window, splice =
       switch(splice,
              fbew = {pGK[i+window-1,1] <- fbewBase*new[length(new)]},
              fbmw = {pGK[i+window-1,1] <- fbewBase*new[length(new)]/new[length(new)-(i+window-1-base)]},
+             wisp = {pGEKS[i+window-1,1] <- splice_t(pGEKS[i+window-2,1], pGEKS[(i-1):(i+window-2)], new, method="window")},
+             hasp = {pGEKS[i+window-1,1] <- splice_t(pGEKS[i+window-2,1], pGEKS[(i-1):(i+window-2)], new, method="half")},
+             mean_pub = {pGEKS[i+window-1,1] <- splice_t(pGEKS[i+window-2,1], pGEKS[(i-1):(i+window-2)], new, method="mean")},
              pGK[i+window-1,1] <- splice_t(pGK[i+window-2,1], old, new, method=splice))
 
     }