+++ R/geks.R
@@ -119,8 +119,8 @@ GEKS_w <- function(x, pvar, qvar, pervar, indexMethod="tornqvist", prodID,
 #' are removed prior to estimating the index for those periods.
 #' @param window An integer specifying the length of the window.
 #' @param splice A character string specifying the splicing method. Valid methods are
-#' window, movement, half, mean, fbew or fbmw. The default is mean. See details for important
-#' considerations when using fbew and fbmw.
+#' window, movement, half, mean, fbew or fbmw, wisp, hasp or mean_pub. The default is mean.
+#' See details for important considerations when using fbew and fbmw.
 #' @param biasAdjust whether to adjust for bias in the coefficients of the bilateral TPD index.
 #' The default is FALSE because making this adjustment will break transitivity of the
 #' GEKS index.
@@ -162,7 +162,7 @@ GEKSIndex <- function(x, pvar, qvar, pervar,indexMethod = "tornqvist", prodID,
   }
 
   # check that only valid splice methods are chosen
-  if(!(tolower(splice) %in% c("mean", "window", "movement", "half", "fbew", "fbmw"))){
+  if(!(tolower(splice) %in% c("mean", "window", "movement", "half", "fbew", "fbmw", "wisp", "hasp", "mean_pub"))){
     stop("Not a valid splicing method.")
   }
 
@@ -254,6 +254,9 @@ GEKSIndex <- function(x, pvar, qvar, pervar,indexMethod = "tornqvist", prodID,
       switch(splice,
              fbew = {pGEKS[i+window-1,1] <- fbewBase*new[length(new)]},
              fbmw = {pGEKS[i+window-1,1] <- fbewBase*new[length(new)]/new[length(new)-(i+window-1-base)]},
+             wisp = {pGEKS[i+window-1,1] <- splice_t(pGEKS[i+window-2,1], pGEKS[(i-1):(i+window-2)], new, method="window")},
+             hasp = {pGEKS[i+window-1,1] <- splice_t(pGEKS[i+window-2,1], pGEKS[(i-1):(i+window-2)], new, method="half")},
+             mean_pub = {pGEKS[i+window-1,1] <- splice_t(pGEKS[i+window-2,1], pGEKS[(i-1):(i+window-2)], new, method="mean")},
              pGEKS[i+window-1,1] <- splice_t(pGEKS[i+window-2,1], old, new, method=splice))
 
     }