+++ news.md
@@ -3,6 +3,7 @@
 ### Improvements
 
 - Added options for the weights when estimating the time-product-dummy index
+- Added the Walsh index as a valid bilateral index to use in GEKSIndex
 
 ### Bug fixes