+++ news.md
@@ -4,7 +4,8 @@
 
 - Added options for the weights when estimating the time-product-dummy index
 - Added the wisp, hasp and mean_pub splicing methods. These are the same methods as window, half and mean splice, but splice onto the published series instead of the previously calculated window. 
-- Added the Walsh and Jevons indexes as bilateral indexes in the GEKSIndex function
+- Added the Walsh and Jevons indexes as bilateral indexes in the GEKSIndex function.
+- Added a function for retrieving data from the Chicago Booth School of Business Dominicks Finer Foods dataset. 
 
 ### Bug fixes