+++ news.md
@@ -10,6 +10,7 @@
 ### Bug fixes
 
 - Fixed issue with the time-product-dummy regression when there are missing products
+- Bug in the FBEW splice for the WTPD multilateral method
 
 
 ## Package update 0.1.3 -> 0.2.0